requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.8
Pillow==9.5.0
celery==5.2.7
redis==4.5.4
requests==2.31.0
mysql-connector-python==8.0.33
gunicorn==20.1.0
psycopg2-binary==2.9.6
